data-pipeline/requirements.txt
requests>=2.31.0
pandas>=2.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
anthropic>=0.40.0
